Complete System Package/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.36.0",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
